# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/auth/merchant/signup": {
      "post": {
        "operationId": "AuthController_signupMerchant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantSignupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailVerificationRequestedDto"
                }
              }
            }
          }
        },
        "summary": "Create a merchant account; emails a verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/merchant/verify-email": {
      "post": {
        "operationId": "AuthController_verifyMerchantEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          }
        },
        "summary": "Confirm the emailed code and sign the new merchant in",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/merchant/verify-email/resend": {
      "post": {
        "operationId": "AuthController_resendMerchantVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendEmailVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestedDto"
                }
              }
            }
          }
        },
        "summary": "Resend the merchant email verification code",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/merchant/login": {
      "post": {
        "operationId": "AuthController_loginMerchant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          }
        },
        "summary": "Log a merchant in with email + password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/staff/login": {
      "post": {
        "operationId": "AuthController_loginStaff",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          }
        },
        "summary": "Log a staff member in with email + password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/customer/otp/request": {
      "post": {
        "operationId": "AuthController_requestCustomerOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestedDto"
                }
              }
            }
          }
        },
        "summary": "Send a login/registration OTP to a customer phone or email",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/customer/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyCustomerOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResultDto"
                }
              }
            }
          }
        },
        "summary": "Verify customer OTP — logs in, or returns a registration token for new phones",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/customer/register": {
      "post": {
        "operationId": "AuthController_registerCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionDto"
                }
              }
            }
          }
        },
        "summary": "Complete customer registration after OTP verification",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensDto"
                }
              }
            }
          }
        },
        "summary": "Rotate a refresh token into a new token pair",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Revoke a refresh token (idempotent)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current actor profile (any role)",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/v1/admin/auth/refresh": {
      "post": {
        "operationId": "AdminAuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Rotate an admin refresh token",
        "tags": [
          "Admin · Auth"
        ]
      }
    },
    "/v1/admin/auth/logout": {
      "post": {
        "operationId": "AdminAuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Revoke the current admin session",
        "tags": [
          "Admin · Auth"
        ]
      }
    },
    "/v1/admin/auth/me": {
      "get": {
        "operationId": "AdminAuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The signed-in admin and what their role permits",
        "tags": [
          "Admin · Auth"
        ]
      }
    },
...
    "/v1/customer/memberships": {
      "post": {
        "operationId": "CustomerPortalController_join",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Join a business loyalty program (idempotent — returns the existing card if already a member)",
        "tags": [
          "Customer"
        ]
      }
    },
    "/v1/customer/cards": {
      "get": {
        "operationId": "CustomerPortalController_cards",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "All loyalty cards belonging to the customer",
        "tags": [
          "Customer"
        ]
      }
    },
    "/v1/customer/cards/{membershipId}": {
      "get": {
        "operationId": "CustomerPortalController_card",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One card with recent stamp activity (poll for live updates)",
        "tags": [
          "Customer"
        ]
      }
    },
    "/v1/customer/cards/{membershipId}/qr": {
      "get": {
        "operationId": "CustomerPortalController_cardQr",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardQrDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The card's code as a QR image — shown at the counter for camera scanning",
        "tags": [
          "Customer"
        ]
      }
    },
...
  },
  "info": {
    "title": "Loyalty Platform API",
    "description": "Multi-tenant loyalty platform — Phase 1. Bearer JWT auth; see /docs for the interactive UI.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "MerchantSignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "owner@brewbean.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "name": {
            "type": "string",
            "example": "Asha Patel"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "EmailVerificationRequestedDto": {
        "type": "object",
        "properties": {
          "verificationRequired": {
            "type": "boolean",
            "example": true,
            "description": "Signup succeeded but the email must be confirmed before sign-in"
          },
          "email": {
            "type": "string",
            "example": "owner@brewbean.com",
            "description": "The email the code was sent to"
          },
          "expiresInSec": {
            "type": "number",
            "example": 900
          },
          "resendInSec": {
            "type": "number",
            "example": 60
          },
          "devCode": {
            "type": "string",
            "description": "Dev-mode only — never present in production"
          }
        },
        "required": [
          "verificationRequired",
          "email",
          "expiresInSec",
          "resendInSec"
        ]
      },
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "owner@brewbean.com"
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
      "TokensDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "accessTokenExpiresInSec": {
            "type": "number",
            "example": 900
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "accessTokenExpiresInSec"
        ]
      },
...
      "AuthSessionDto": {
        "type": "object",
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/TokensDto"
          },
          "actor": {
            "$ref": "#/components/schemas/SessionActorDto"
          },
          "business": {
            "nullable": true,
            "description": "The actor’s business (merchant: own business or null; staff: employer; customer: null)",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessDto"
              }
            ]
          }
        },
        "required": [
          "tokens",
          "actor",
          "business"
        ]
      },
      "ResendEmailVerificationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "owner@brewbean.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "OtpRequestedDto": {
        "type": "object",
        "properties": {
          "expiresInSec": {
            "type": "number",
            "example": 300
          },
          "resendInSec": {
            "type": "number",
            "example": 60
          },
          "devCode": {
            "type": "string",
            "description": "Dev-mode only — never present in production"
          }
        },
        "required": [
          "expiresInSec",
          "resendInSec"
        ]
      },
      "EmailLoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "owner@brewbean.com"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "+91 98765 00001",
            "description": "Phone number or email address. Phones accept any format and are normalised to E.164 server-side; emails are lowercased. Customers may use either — staff and merchants must use a phone here, since they already have email+password sign-in."
          }
        },
        "required": [
          "identifier"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "+91 98765 00001",
            "description": "Phone number or email address. Phones accept any format and are normalised to E.164 server-side; emails are lowercased. Customers may use either — staff and merchants must use a phone here, since they already have email+password sign-in."
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "identifier",
          "code"
        ]
      },
      "AuthResultDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "AUTHENTICATED",
              "REGISTRATION_REQUIRED"
            ]
          },
          "session": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthSessionDto"
              }
            ]
          },
          "registrationToken": {
            "type": "string",
            "nullable": true,
            "description": "Present when status is REGISTRATION_REQUIRED; pass to /register within 10 minutes"
          }
        },
        "required": [
          "status",
          "session",
          "registrationToken"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "registrationToken": {
            "type": "string",
            "description": "registrationToken returned by the verify step"
          },
          "name": {
            "type": "string",
            "example": "Asha Patel"
          }
        },
        "required": [
          "registrationToken",
          "name"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "MeDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "MERCHANT",
              "STAFF",
              "CUSTOMER"
            ]
          },
          "actor": {
            "$ref": "#/components/schemas/SessionActorDto"
          },
          "business": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/BusinessDto"
              }
            ]
          },
          "impersonation": {
            "nullable": true,
            "description": "Set when this session is an admin impersonating the merchant",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ImpersonationInfoDto"
              }
            ]
          }
        },
        "required": [
          "role",
          "actor",
          "business",
          "impersonation"
        ]
      },
...
      "AdminProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "OPS",
              "SUPPORT",
              "FINANCE",
              "ANALYST"
            ]
          },
          "capabilities": {
            "description": "What this role may do",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "recoveryCodesRemaining": {
            "type": "number"
          },
          "lastLoginAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "capabilities",
          "twoFactorEnabled",
          "recoveryCodesRemaining",
          "lastLoginAt",
          "createdAt"
        ]
      },
...
      "AdminRefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "JoinBusinessDto": {
        "type": "object",
        "properties": {
          "businessSlug": {
            "type": "string",
            "example": "brew-and-bean",
            "description": "Business slug from the QR join link"
          },
          "marketingConsent": {
            "type": "boolean",
            "description": "Marketing opt-in. Recorded with the exact wording shown at enrolment."
          }
        },
        "required": [
          "businessSlug"
        ]
      },
...
      "CardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Membership id — the canonical card identifier"
          },
          "code": {
            "type": "string",
            "example": "7F3K9QZP",
            "description": "Unique customer code (stored form)"
          },
          "formattedCode": {
            "type": "string",
            "example": "7F3K-9QZP",
            "description": "Display form of the code"
          },
          "stampCount": {
            "type": "number",
            "example": 4,
            "description": "Stamps on the current card"
          },
          "completedCount": {
            "type": "number",
            "example": 1,
            "description": "Rewards earned (completed cards)"
          },
          "totalStamps": {
            "type": "number",
            "example": 14,
            "description": "Lifetime stamps"
          },
          "lastStampAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pendingRewards": {
            "description": "Rewards earned but not yet handed over",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RedemptionSummaryDto"
            }
          },
          "redeemedCount": {
            "type": "number",
            "example": 1,
            "description": "Rewards already honoured (completedCount − pending)"
          },
          "business": {
            "$ref": "#/components/schemas/CardBusinessDto"
          },
          "campaign": {
            "$ref": "#/components/schemas/CardCampaignDto"
          },
          "style": {
            "$ref": "#/components/schemas/CardStyleDto"
          }
        },
        "required": [
          "id",
          "code",
          "formattedCode",
          "stampCount",
          "completedCount",
          "totalStamps",
          "lastStampAt",
          "joinedAt",
          "pendingRewards",
          "redeemedCount",
          "business",
          "campaign",
          "style"
        ]
      },
      "JoinResultDto": {
        "type": "object",
        "properties": {
          "card": {
            "$ref": "#/components/schemas/CardDto"
          },
          "alreadyMember": {
            "type": "boolean",
            "description": "True when the customer already had this card"
          }
        },
        "required": [
          "card",
          "alreadyMember"
        ]
      },
...
    }
  }
}